Change to eclipse project
