+++ org.eclipse.emfcloud.validation.framework/.classpath
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
+		<attributes>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
+	<classpathentry kind="src" path="src/main/java/"/>
+	<classpathentry kind="src" path="src/test/java/"/>
+	<classpathentry kind="src" path="target/generated-sources/annotations/"/>
+	<classpathentry kind="src" path="target/generated-test-sources/test-annotations/"/>
+	<classpathentry kind="output" path="target/classes"/>
+</classpath>
