+++ org.eclipse.emfcloud.validation.framework/META-INF/MANIFEST.MF
@@ -1,0 +1,40 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Modelserver Validation
+Bundle-SymbolicName: org.eclipse.emfcloud.validation.framework
+Bundle-Version: 1.0.0.qualifier
+Export-Package: 
+ org.eclipse.emfcloud.validation.framework
+Require-Bundle: org.apache.log4j;bundle-version="1.2.15",
+ com.fasterxml.jackson.core.jackson-annotations;bundle-version="2.10.3",
+ com.fasterxml.jackson.core.jackson-core;bundle-version="2.10.3",
+ com.fasterxml.jackson.core.jackson-databind;bundle-version="2.10.3",
+ org.eclipse.emfcloud.modelserver.client;bundle-version="0.7.0",
+ org.emfjson.jackson;bundle-version="1.2.0",
+ org.junit;bundle-version="4.13.0",
+ org.eclipse.emfcloud.modelserver.emf;bundle-version="0.7.0",
+ org.eclipse.emf.common;bundle-version="2.20.0",
+ org.eclipse.emf.ecore;bundle-version="2.23.0",
+ org.eclipse.emf.ecore.xmi;bundle-version="2.16.0",
+ org.mockito;bundle-version="2.23.0",
+ org.eclipse.emfcloud.modelserver.common;bundle-version="0.7.0",
+ org.eclipse.emfcloud.modelserver.lib
+Bundle-ClassPath: .
+Import-Package: okhttp3,
+ okhttp3.internal,
+ okhttp3.internal.authenticator,
+ okhttp3.internal.cache,
+ okhttp3.internal.cache2,
+ okhttp3.internal.concurrent,
+ okhttp3.internal.connection,
+ okhttp3.internal.http,
+ okhttp3.internal.http1,
+ okhttp3.internal.http2,
+ okhttp3.internal.io,
+ okhttp3.internal.platform,
+ okhttp3.internal.platform.android,
+ okhttp3.internal.proxy,
+ okhttp3.internal.publicsuffix,
+ okhttp3.internal.tls,
+ okhttp3.internal.ws
+Bundle-Vendor: EclipseSource
