+++ org.eclipse.emfcloud.validation.framework/build.properties
@@ -1,0 +1,6 @@
+source.. = src/main/java/,\
+           src/test/java/,\
+           target/generated-sources/annotations/,\
+           target/generated-test-sources/test-annotations/
+bin.includes = META-INF/,\
+               .
