+++ org.eclipse.emfcloud.validation.framework/src/main/java/org/eclipse/emfcloud/validation/framework/ValidationResultChangeListener.java
@@ -1,0 +1,19 @@
+/*******************************************************************************
+ * Copyright (c) 2020-2021 EclipseSource and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * https://www.eclipse.org/legal/epl-2.0, or the MIT License which is
+ * available at https://opensource.org/licenses/MIT.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR MIT
+ ******************************************************************************/
+package org.eclipse.emfcloud.validation.framework;
+
+import java.util.List;
+
+public abstract class ValidationResultChangeListener {
+
+	public abstract void changed(List<ValidationResult> newResult);
+
+}
