+++ org.eclipse.emfcloud.validation.framework/src/test/java/org/eclipse/emfcloud/validation/framework/emptyTest.ecore
