+++ releng/org.eclipse.emfcloud.validation.codestyle/pom.xml
@@ -1,0 +1,13 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<artifactId>org.eclipse.emfcloud.validation.codestyle</artifactId>
+	<parent>
+		<groupId>org.eclipse.emfcloud.validation</groupId>
+		<artifactId>org.eclipse.emfcloud.validation.releng</artifactId>
+		<relativePath>../pom.xml</relativePath>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+	<packaging>pom</packaging>
+</project>
