+++ releng/org.eclipse.emfcloud.validation.repository/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>model-validation</name>
+	<name>org.eclipse.emfcloud.validation.repository</name>
 	<comment></comment>
 	<projects>
 	</projects>
