+++ targetplatform/pom.xml
@@ -1,0 +1,18 @@
+<!-- Copyright (C) 2011, Obeo All rights reserved. This program and the accompanying 
+	materials are made available under the terms of the Eclipse Public License 
+	v1.0 which accompanies this distribution, and is available at http://www.eclipse.org/legal/epl-v10.html -->
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>org.eclipse.emfcloud.validation</groupId>
+	<artifactId>targetplatform</artifactId>
+	<version>1.0.0-SNAPSHOT</version>
+	<packaging>pom</packaging>
+	<properties>
+		<tycho-version>2.1.0</tycho-version>
+		<tycho-extras-version>2.1.0</tycho-extras-version>
+		<targetPlatform.application.name>org.eclipse.cbi.targetplatform.tpd.converter</targetPlatform.application.name>
+		<targetPlatform.application.args>r2020-06.tpd</targetPlatform.application.args>
+	</properties>
+</project>
